Move membership schema to own file
