+++ packages/core/src/modules/user/user-service.ts
@@ -1,4 +1,5 @@
 import type {
+  Membership,
   NotificationPermissions,
   NotificationPermissionsWrite,
   PrivacyPermissions,
