+++ packages/types/src/index.ts
@@ -5,6 +5,8 @@ export * from "./attendance/attendee"
 export * from "./attendance/attendance-selections"
 export * from "./attendance/waitlist-attendee"
 export * from "./company"
+export * from "./user/user"
+export * from "./user/membership"
 export * from "./event"
 export * from "./event-hosting-group"
 export * from "./group"
@@ -17,4 +19,3 @@ export * from "./offline"
 export * from "./payment"
 export * from "./personal-mark"
 export * from "./privacy-permissions"
-export * from "./user"
